+++ src/main/java/team/cutano/swiftmessengerservice/pojo/User.java
@@ -1,23 +1,23 @@
 package team.cutano.swiftmessengerservice.pojo;
 
 public class User {
-    private Long userID;
+    private Integer userID;
     private String username;
     private String avatar;
     private String password;
 
-    public User(Long userID, String username, String avatar, String password) {
+    public User(Integer userID, String username, String avatar, String password) {
         this.userID = userID;
         this.username = username;
         this.avatar = avatar;
         this.password = password;
     }
 
-    public Long getUserID() {
+    public Integer getUserID() {
         return userID;
     }
 
-    public void setUserID(Long userID) {
+    public void setUserID(Integer userID) {
         this.userID = userID;
     }
